% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonPlot.R
\name{seasonPlot}
\alias{seasonPlot}
\title{seasonPlot: create seasonality plot for stock prices or cryptocurrencies}
\usage{
seasonPlot(
  Symbols,
  StartYear = 2000,
  EndYear = 2020,
  LineColor = 1,
  xlab = "Monthly",
  BackgroundMode = TRUE,
  alpha = 0.05,
  OutpuData = FALSE,
  Save = FALSE,
  output_width = 1000,
  output_height = 700,
  family = "Helvetica"
)
}
\arguments{
\item{Symbols}{a character vector specifying the names of each symbol to be loaded}

\item{StartYear}{a numeric of start year (Common Er)}

\item{EndYear}{a numeric of end year (Common Er)}

\item{LineColor}{a numeric; if the value is 1, red1 is selected,
if the value is 2, blue1 is selected,
if the value is 3, green1 is selected, and
if the value is 4, black is selected.
If BackgroundMode is TRUE, this argument is disabled.}

\item{xlab}{a character of X-axis label.}

\item{BackgroundMode}{a logical; draw a baclground color by react.}

\item{alpha}{a numeric; The alpha parameter is a number between 0.0 (fully transparent) and 1.0 (fully opaque).}

\item{OutpuData}{a logical; output as a data.frame or not.}

\item{Save}{a logical; save as image or not}

\item{output_width}{a output size of width (pixel). Initial value recommended.}

\item{output_height}{a output size of height (pixel). Initial value recommended.}

\item{family}{a charactor of font.}
}
\value{
plot result
}
\description{
This function is to create a seasonality plot with some color options.
It uses the same symbols as the quantmod package (e.g. SPY, BTC-USD, and ETH-USD etc).
}
\examples{
\donttest{

# Search by SPY (S&P500 index)
#seasonPlot(Symbols = "SPY")

}

}
\author{
Satoshi Kume
}
